+++ lib/avro_turf/in_memory_cache.rb
@@ -5,6 +5,7 @@ class AvroTurf::InMemoryCache
   def initialize
     @schemas_by_id = {}
     @ids_by_schema = {}
+    @schema_by_subject_version = {}
   end
 
   def lookup_by_id(id)
@@ -16,12 +17,22 @@ class AvroTurf::InMemoryCache
   end
 
   def lookup_by_schema(subject, schema)
-    key = subject + schema.to_s
+    key = [subject, schema.to_s]
     @ids_by_schema[key]
   end
 
   def store_by_schema(subject, schema, id)
-    key = subject + schema.to_s
+    key = [subject, schema.to_s]
     @ids_by_schema[key] = id
+  end
+
+  def lookup_by_version(subject, version)
+    key = "#{subject}#{version}"
+    @schema_by_subject_version[key]
+  end
+
+  def store_by_version(subject, version, schema)
+    key = "#{subject}#{version}"
+    @schema_by_subject_version[key] = schema
   end
 end